Add jdk17 variants
These should be the new default (since 17 is LTS), but the longevity of vanilla builds from http://jdk.java.net/17/ are not necessarily a given (Oracle usually only supplies binaries through that site for ~6 months IIRC).

So, until that is confirmed to be a long-term maintained home for new 17 builds or 17 shows up somewhere like https://github.com/adoptopenjdk?q=upstream, we should stick to 11 as our default. 😞

Co-authored-by: denistazhkenov <denis.tazhkenov@careem.com>